recaptcha added

+++ shop/views.py
@@ -7,25 +7,29 @@ from django.core.files import File
 from django.shortcuts import render
 
 from BotInterface import BotInterface
-from .models import Category, Product, ProductImage, Review, ProductCharacteristic, ReviewImages, ProductOption, Basket, ProductInBasket
+from utils.recaptcha import checkReCAPTHA
+from .models import Category, Product, ProductImage, Review, ProductCharacteristic, ReviewImages, ProductOption, Basket, \
+    ProductInBasket
 from CRM.models import Order, ProductInOrder
 
 
 # Create your views here.
 def index(request):
     hits = Product.objects.filter(inactive=False).order_by('-buy_count')[:10]
-    categories = Category.objects.all() 
-    return render(request, 'shop/shop.html', {"hits" : hits, "categories":categories})
+    categories = Category.objects.all()
+    return render(request, 'shop/shop.html', {"hits": hits, "categories": categories})
+
 
 def category(request, category_id):
     try:
         category = Category.objects.get(id=category_id)
         products = Product.objects.filter(categories__id=category_id).filter(inactive=False)
-        response = render(request, 'shop/category.html', {"products":products,"category":category})
+        response = render(request, 'shop/category.html', {"products": products, "category": category})
         return response
     except:
         return render(request, 'shop/notfound.html')
-    
+
+
 def product(request, product_id):
     try:
         review = Review.objects.get(id=request.GET["review_id"])
@@ -62,21 +66,30 @@ def product(request, product_id):
         options_founded = True
     else:
         options_founded = False
-    return render(request, 'shop/product.html', {"product":product,"product_images":product_images, "characteristics":characteristics,"reviews":reviews, "product_images_founded":product_images_founded, "characteristics_founded":characteristics_founded, "reviews_founded":reviews_founded, "reviews_more_than_2":reviews_more_than_2, "review":review, "options_founded":options_founded, "options":options})
+    return render(request, 'shop/product.html',
+                  {"product": product, "product_images": product_images, "characteristics": characteristics,
+                   "reviews": reviews, "product_images_founded": product_images_founded,
+                   "characteristics_founded": characteristics_founded, "reviews_founded": reviews_founded,
+                   "reviews_more_than_2": reviews_more_than_2, "review": review, "options_founded": options_founded,
+                   "options": options})
+
 
 def search(request):
     query = request.GET.get('query')
-    
+
     if not query:
         return redirect(request.META.get('HTTP_REFERER'))
-    
+
     products = Product.objects.filter(articul=query)
     if len(products) == 0:
-        products = products | Product.objects.filter(name__startswith=query.lower()) |  Product.objects.filter(name__startswith=query.upper()) | Product.objects.filter(name__startswith=query.capitalize())  
+        products = products | Product.objects.filter(name__startswith=query.lower()) | Product.objects.filter(
+            name__startswith=query.upper()) | Product.objects.filter(name__startswith=query.capitalize())
     if len(products) == 0:
-        products = products | Product.objects.filter(name__contains=query.lower()) |  Product.objects.filter(name__contains=query.upper()) | Product.objects.filter(name__contains=query.capitalize())  
+        products = products | Product.objects.filter(name__contains=query.lower()) | Product.objects.filter(
+            name__contains=query.upper()) | Product.objects.filter(name__contains=query.capitalize())
     if len(products) == 0:
-        products = products | Product.objects.filter(des__contains=query.lower())  | Product.objects.filter(des__contains=query.upper()) | Product.objects.filter(des__contains=query.capitalize())
+        products = products | Product.objects.filter(des__contains=query.lower()) | Product.objects.filter(
+            des__contains=query.upper()) | Product.objects.filter(des__contains=query.capitalize())
     if len(products) >= 20:
         products = products[0:20]
 
@@ -85,8 +98,9 @@ def search(request):
         founded = False
 
     products = products.filter(inactive=False)
-        
-    return render(request, 'shop/search.html', {"products":products,"founded":founded,"query":query})
+
+    return render(request, 'shop/search.html', {"products": products, "founded": founded, "query": query})
+
 
 @csrf_exempt
 def search_recomendations(request):
@@ -94,11 +108,14 @@ def search_recomendations(request):
         query = request.POST["query"]
         products = Product.objects.filter(articul=query)
         if len(products) == 0:
-            products = products | Product.objects.filter(name__startswith=query.lower()) |  Product.objects.filter(name__startswith=query.upper()) | Product.objects.filter(name__startswith=query.capitalize())  
+            products = products | Product.objects.filter(name__startswith=query.lower()) | Product.objects.filter(
+                name__startswith=query.upper()) | Product.objects.filter(name__startswith=query.capitalize())
         if len(products) == 0:
-            products = products | Product.objects.filter(name__contains=query.lower()) |  Product.objects.filter(name__contains=query.upper()) | Product.objects.filter(name__contains=query.capitalize())  
+            products = products | Product.objects.filter(name__contains=query.lower()) | Product.objects.filter(
+                name__contains=query.upper()) | Product.objects.filter(name__contains=query.capitalize())
         if len(products) == 0:
-            products = products | Product.objects.filter(des__contains=query.lower())  | Product.objects.filter(des__contains=query.upper()) | Product.objects.filter(des__contains=query.capitalize())
+            products = products | Product.objects.filter(des__contains=query.lower()) | Product.objects.filter(
+                des__contains=query.upper()) | Product.objects.filter(des__contains=query.capitalize())
         if len(products) >= 10:
             products = products[0:10]
 
@@ -111,10 +128,10 @@ def get_reviews(request):
     if request.method == 'GET':
         reviews = list(Review.objects.filter(product=request.GET['id']).values())
         for i in range(len(reviews)):
-            reviews[i]["created_at"] = defaultfilters.date(reviews[i]["created_at"]) + ' ' + defaultfilters.time(reviews[i]["created_at"])
+            reviews[i]["created_at"] = defaultfilters.date(reviews[i]["created_at"]) + ' ' + defaultfilters.time(
+                reviews[i]["created_at"])
         return JsonResponse({'reviews': reviews})
     return JsonResponse({'status': 'Invalid request'}, status=400)
-    
 
 
 def basket(request):
@@ -125,32 +142,37 @@ def basket(request):
             product.delete()
     return render(request, 'shop/basket.html', {})
 
+
 def create_order(request):
-    basket = Basket.objects.get(unique_id=request.COOKIES.get('basket_uid'))
-    order = Order()
-    order.contacts = request.POST.get('contacts')
-    order.summ = request.POST.get('successful_basket_summ')
-    
-    products = ProductInBasket.objects.filter(basket=basket)
-    if len(products) == 0:
-        return render(request, 'CRM/message.html', {"text":"Не удалось сформировать заказ. Корзина пуста или заказ уже находится в обработке"})
-    order.save()
-    products_in_order = []
-    for product in products:
-        p = ProductInOrder()
-        p.count = product.count
-        p.product = product.product
-        p.order = order
-        p.options = product.options
-        p.save()
-        product.delete()
-        products_in_order.append(p)
+    if request.method == 'POST':
+        if not checkReCAPTHA(request):
+            return render(request, 'CRM/message.html',
+                          {"text": "Вы не прошли проверку на робота. Попробуйте еще раз."})
+        basket = Basket.objects.get(unique_id=request.COOKIES.get('basket_uid'))
+        order = Order()
+        order.contacts = request.POST.get('contacts')
+        order.summ = request.POST.get('successful_basket_summ')
 
-    BotInterface.create_order(order, products_in_order)
+        products = ProductInBasket.objects.filter(basket=basket)
+        if len(products) == 0:
+            return render(request, 'CRM/message.html',
+                          {"text": "Не удалось сформировать заказ. Корзина пуста или заказ уже находится в обработке"})
+        order.save()
+        products_in_order = []
+        for product in products:
+            p = ProductInOrder()
+            p.count = product.count
+            p.product = product.product
+            p.order = order
+            p.options = product.options
+            p.save()
+            product.delete()
+            products_in_order.append(p)
 
-    return render(request, 'CRM/message.html', {"text": "Ваш заказ уже принят в обработку. Скоро с вами свяжется менеджер"})
+        BotInterface.create_order(order, products_in_order)
 
-
+        return render(request, 'CRM/message.html',
+                      {"text": "Ваш заказ уже принят в обработку. Скоро с вами свяжется менеджер"})
 
 
 @csrf_exempt
@@ -169,8 +191,9 @@ def save_image_for_review(request):
             review_img.img = File(file, name=file.name)
             review_img.save()
             imgs_url.append(review_img.img.url)
-        return JsonResponse({'review_id':review.id,'imgs_url':imgs_url})
+        return JsonResponse({'review_id': review.id, 'imgs_url': imgs_url})
     return JsonResponse({'status': 'Invalid request'}, status=400)
+
 
 def save_review(request):
     if request.method == 'POST':
@@ -197,38 +220,38 @@ def save_review(request):
             return redirect(f"{request.META.get('HTTP_REFERER')}?review_id={review.id}")
         else:
             return redirect(f'/shop/products/{review.product.id}')
-            
 
 
 @csrf_exempt
 def put_in_basket(request):
     if request.method == 'POST':
-        basket = Basket.objects.get(unique_id=request.COOKIES.get('basket_uid'))
-        product = Product.objects.get(id=request.POST["product_id"])
+        basket_ = Basket.objects.get(unique_id=request.COOKIES.get('basket_uid'))
+        product_ = Product.objects.get(id=request.POST["product_id"])
         if request.POST.get('option_id'):
             option = ProductOption.objects.get(id=request.POST["option_id"])
         else:
-            options = ProductOption.objects.filter(product=product)
+            options = ProductOption.objects.filter(product=product_)
             if options:
                 option = options[0]
             else:
                 option = 0
         try:
-            del_product = ProductInBasket.objects.get(basket = basket, product = product)
+            del_product = ProductInBasket.objects.get(basket=basket_, product=product_)
             del_product.delete()
-            return JsonResponse({"status":"deleted"})
+            return JsonResponse({"status": "deleted"})
         except:
             p = ProductInBasket()
             p.count = request.POST["count"]
-            p.product = product
-            p.basket = basket
+            p.product = product_
+            p.basket = basket_
             if option:
                 p.options = option
             else:
                 p.options = None
             p.save()
-            return JsonResponse({'status':'added'})
-        
+            return JsonResponse({'status': 'added'})
+
+
 @csrf_exempt
 def update_in_basket(request):
     if request.method == 'POST':
@@ -239,27 +262,29 @@ def update_in_basket(request):
         else:
             option = 0
         try:
-            upd_product = ProductInBasket.objects.get(basket = basket, product = product)
+            upd_product = ProductInBasket.objects.get(basket=basket, product=product)
             upd_product.count = request.POST["count"]
             if option:
                 upd_product.options = option
             upd_product.save()
-            return JsonResponse({'status':'updated'})
+            return JsonResponse({'status': 'updated'})
         except:
-            return JsonResponse({'status':'not updated'})
-        
+            return JsonResponse({'status': 'not updated'})
+
+
 @csrf_exempt
 def delete_in_basket(request):
     if request.method == 'POST':
         basket = Basket.objects.get(unique_id=request.COOKIES.get('basket_uid'))
         product = Product.objects.get(id=request.POST["product_id"])
         try:
-            del_product = ProductInBasket.objects.get(basket = basket, product = product)
+            del_product = ProductInBasket.objects.get(basket=basket, product=product)
             del_product.delete()
-            return JsonResponse({"status":"deleted"})
+            return JsonResponse({"status": "deleted"})
         except:
-            return JsonResponse({"status":"not deleted"})
-        
+            return JsonResponse({"status": "not deleted"})
+
+
 @csrf_exempt
 def get_products_in_basket_count(request):
     basket = Basket.objects.get(unique_id=request.COOKIES.get('basket_uid'))
